Added ReceiptControlListner
set up interface for receipt listener
